Re-arrange files in channel-client
