+++ packages/channel-client/src/error.ts
@@ -1,0 +1,7 @@
+export enum ErrorCodes {
+  SIGNING_ADDRESS_NOT_FOUND = 1000,
+  INVALID_APP_DEFINITION = 1001,
+  INVALID_APP_DATA = 1002,
+  UNSUPPORTED_TOKEN = 1003,
+  CHANNEL_NOT_FOUND = 1004
+}
